+++ plugins/card-assets/lang/cs.json
@@ -41,7 +41,6 @@
     "NoChildren": "Žádné děti",
     "ConfigDescription": "Rozšíření pro správu znalostí ve stylu databáze.",
     "AddCollaborators": "Přidat spolupracovníky",
-    "Home": "Domů",
     "CardTitle": "Název",
     "Post": "Přidat",
     "Compact": "Kompaktní",
