+++ plugins/card-assets/lang/de.json
@@ -41,7 +41,6 @@
     "NoChildren": "Keine Kinder",
     "ConfigDescription": "Erweiterung für Datenbank-basiertes Wissensmanagement.",
     "AddCollaborators": "Mitarbeiter hinzufügen",
-    "Home": "Startseite",
     "CardTitle": "Titel",
     "Post": "Beitrag",
     "Compact": "Kompakt",
