+++ plugins/card-assets/lang/en.json
@@ -41,7 +41,6 @@
     "NoChildren": "No children",
     "ConfigDescription": "Extension for database-style knowledge management.",
     "AddCollaborators": "Add collaborators",
-    "Home": "Home",
     "CardTitle": "Title",
     "Post": "Post",
     "Compact": "Compact",
