+++ plugins/card-assets/lang/es.json
@@ -41,7 +41,6 @@
     "NoChildren": "No hay hijos",
     "ConfigDescription": "Extensión para la gestión del conocimiento al estilo de una base de datos.",
     "AddCollaborators": "Agregar colaboradores",
-    "Home": "Inicio",
     "CardTitle": "Título",
     "Post": "Publicar",
     "Compact": "Compacto",
