+++ plugins/card-assets/lang/fr.json
@@ -41,7 +41,6 @@
     "NoChildren": "Pas d'enfants",
     "ConfigDescription": "Extension pour la gestion des connaissances de style base de données.",
     "AddCollaborators": "Ajouter des collaborateurs",
-    "Home": "Accueil",
     "CardTitle": "Titre",
     "Post": "Publier",
     "Compact": "Compact",
