+++ plugins/card-assets/lang/it.json
@@ -41,7 +41,6 @@
     "NumberTypes": "{count, plural, one {# tipo} other {# tipi}}",
     "ConfigDescription": "Estensione per la gestione della conoscenza in stile database.",
     "AddCollaborators": "Aggiungi collaboratori",
-    "Home": "Home",
     "CardTitle": "Titolo",
     "Post": "Pubblica",
     "Compact": "Compatto",
