+++ plugins/card-assets/lang/ja.json
@@ -41,7 +41,6 @@
     "NoChildren": "子がありません",
     "ConfigDescription": "データベーススタイルのナレッジマネジメント用の拡張機能。",
     "AddCollaborators": "コラボレーターを追加",
-    "Home": "ホーム",
     "CardTitle": "タイトル",
     "Post": "投稿",
     "Compact": "コンパクト",
