+++ plugins/card-assets/lang/pt.json
@@ -41,7 +41,6 @@
     "NoChildren": "Sem filhos",
     "ConfigDescription": "Extensão para gerenciamento de conhecimento no estilo de banco de dados.",
     "AddCollaborators": "Adicionar colaboradores",
-    "Home": "Início",
     "CardTitle": "Título",
     "Post": "Publicar",
     "Compact": "Compacto",
