+++ plugins/card-assets/lang/ru.json
@@ -41,7 +41,6 @@
     "NoChildren": "Нет потомков",
     "ConfigDescription": "Расширение для управления знаниями в стиле базы данных.",
     "AddCollaborators": "Добавить сотрудников",
-    "Home": "Дом",
     "CardTitle": "Название",
     "Post": "Опубликовать",
     "Compact": "Компактный",
