+++ plugins/card-assets/lang/zh.json
@@ -41,7 +41,6 @@
     "NoChildren": "没有子级",
     "ConfigDescription": "用于数据库风格知识管理的扩展。",
     "AddCollaborators": "添加协作者",
-    "Home": "主页",
     "CardTitle": "标题",
     "Post": "发布",
     "Compact": "紧凑",
