+++ plugins/card-resources/src/types.ts
@@ -18,7 +18,6 @@ import { type Heading } from '@hcengineering/text-editor'
 
 interface BaseNavigatorConfig {
   types: Array<Ref<MasterTag>>
-  home?: boolean
   groupBySpace?: boolean
   savedViews?: boolean
   allowCreate?: boolean
