+++ plugins/home-assets/lang/cs.json
@@ -3,6 +3,9 @@
     "Home": "Domov",
     "All": "Všechny aktualizace",
     "CreatedByMe": "Vytvořeno mnou",
-    "Threads": "Vlákna"
+    "Threads": "Vlákna",
+    "Compact": "Kompaktní",
+    "Comfortable": "Komfortní",
+    "Comfortable2": "Komfortní 2"
   }
 }
