+++ plugins/home-assets/lang/de.json
@@ -3,6 +3,9 @@
     "Home": "Startseite",
     "All": "Alle Updates",
     "CreatedByMe": "Von mir erstellt",
-    "Threads": "Threads"
+    "Threads": "Threads",
+    "Compact": "Kompakt",
+    "Comfortable": "Komfortabel",
+    "Comfortable2": "Komfortabel 2"
   }
 }
