+++ plugins/home-assets/lang/en.json
@@ -3,6 +3,9 @@
     "Home": "Home",
     "All": "All updates",
     "CreatedByMe": "Created by me",
-    "Threads": "Threads"
+    "Threads": "Threads",
+    "Compact": "Compact",
+    "Comfortable": "Comfortable",
+    "Comfortable2": "Comfortable 2"
   }
 }
