+++ plugins/home-assets/lang/es.json
@@ -3,6 +3,9 @@
     "Home": "Inicio",
     "All": "Todas las actualizaciones",
     "CreatedByMe": "Creado por mí",
-    "Threads": "Hilos"
+    "Threads": "Hilos",
+    "Compact": "Compacto",
+    "Comfortable": "Confortable",
+    "Comfortable2": "Confortable 2"
   }
 }
