+++ plugins/home-assets/lang/fr.json
@@ -3,6 +3,9 @@
     "Home": "Accueil",
     "All": "Toutes les mises à jour",
     "CreatedByMe": "Créé par moi",
-    "Threads": "Fils"
+    "Threads": "Fils",
+    "Compact": "Compact",
+    "Comfortable": "Confortable",
+    "Comfortable2": "Confortable 2"
   }
 }
