+++ plugins/home-assets/lang/it.json
@@ -3,6 +3,9 @@
     "Home": "Casa",
     "All": "Tutti gli aggiornamenti",
     "CreatedByMe": "Creato da me",
-    "Threads": "Thread"
+    "Threads": "Thread",
+    "Compact": "Compatto",
+    "Comfortable": "Comodo",
+    "Comfortable2": "Comodo 2"
   }
 }
