+++ plugins/home-assets/lang/pt.json
@@ -3,6 +3,9 @@
     "Home": "Início",
     "All": "Todas as atualizações",
     "CreatedByMe": "Criado por mim",
-    "Threads": "Tópicos"
+    "Threads": "Threads",
+    "Compact": "Compacto",
+    "Comfortable": "Confortável",
+    "Comfortable2": "Confortável 2"
   }
 }
