+++ plugins/home-assets/lang/ru.json
@@ -3,6 +3,9 @@
     "Home": "Главная",
     "All": "Все обновления",
     "CreatedByMe": "Созданные мной",
-    "Threads": "Треды"
+    "Threads": "Потоки",
+    "Compact": "Компактный",
+    "Comfortable": "Комфортный",
+    "Comfortable2": "Комфортный 2"
   }
 }
