+++ plugins/home-resources/package.json
@@ -41,7 +41,12 @@
     "@hcengineering/analytics": "^0.6.0",
     "@hcengineering/card": "^0.6.0",
     "@hcengineering/card-resources": "^0.6.0",
+    "@hcengineering/chat": "^0.6.0",
     "@hcengineering/core": "^0.6.32",
+    "@hcengineering/contact": "^0.6.24",
+    "@hcengineering/contact-resources": "^0.6.0",
+    "@hcengineering/communication-types": "^0.1.0",
+    "@hcengineering/communication-resources": "^0.6.0",
     "@hcengineering/home": "^0.6.0",
     "@hcengineering/platform": "^0.6.11",
     "@hcengineering/presentation": "^0.6.3",
